cleanup WithNavToNearestAncestor mixin
move the code to the single controller using it
